+++ include/reference_config_templates/Mus_musculus/GENCODE_M25.yaml
@@ -7,9 +7,12 @@ species: "Mus musculus"
 
 genome:
   url: "https://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_mouse/release_M25/GRCm38.primary_assembly.genome.fa.gz"
+  postprocess: lib.postprocess.default
+
 
 annotation:
   url: "https://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_mouse/release_M25/gencode.vM25.primary_assembly.annotation.gtf.gz"
+  postprocess: lib.postprocess.default
   
 rrna:
   url:
